+++ client/src/input/Block.js
@@ -736,9 +736,16 @@ export default Garnish.Base.extend({
 
   saveExpansion () {
     if (!this.isNew()) {
+      // Use the duplicated block ID if we're on a new provisional draft
+      // The server-side code will also apply the new state to the canonical block
+      const thisBlockId = this.getId()
+      const duplicatedBlockId = window.draftEditor.duplicatedElements[thisBlockId]
+      const sentBlockId = window.draftEditor.settings.isProvisionalDraft && typeof duplicatedBlockId !== 'undefined'
+        ? duplicatedBlockId
+        : thisBlockId
       Craft.queueActionRequest('neo/input/save-expansion', {
         expanded: this.isExpanded() ? 1 : 0,
-        blockId: this.getId(),
+        blockId: sentBlockId,
         locale: this.getLocale()
       })
     }