+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 - Fixed a bug where GraphQL queries wouldn't return Neo child blocks in some cases
 - Fixed a bug where an entry's provisional draft would be created or updated if a Neo block was expanded or collapsed
+- Fixed a bug where expanding or collapsing a Neo block on a provisional draft wouldn't also set the expanded/collapsed state on the canonical block
 
 ## 2.11.17 - 2021-11-11